feat(Demo) gather of datepicker and header example
