+++ packages/react-vapor/src/components/datePicker/examples/DatePickerExamples.1.use.md
@@ -1,0 +1,13 @@
+## Guidelines
+
+Whenever possible, provide ready-made selections of the most contextually relevant values (e.g., last week, last month, etc.).
+
+## Variations
+
+The date picker type choice depends on the input to submit and on the precision required.
+
+| Type                | Purpose                                                                                                                       |
+| ------------------- | ----------------------------------------------------------------------------------------------------------------------------- |
+| Single Date         | Used to select a specific moment.                                                                                             |
+| Date Range          | Used to select a period of time, ranging from one to multiple days.                                                           |
+| With Time Selection | Single Date and Date Range pickers can both be configured to allow for hour and minute selection on each date to be selected. |
