--- packages/react-vapor/src/components/headers/examples/BreadcrumbHeaderExample.tsx
@@ -1,112 +1,0 @@
-import * as React from 'react';
-import {Button} from '../../button/Button';
-import {Svg} from '../../svg/Svg';
-import {BreadcrumbHeader} from '../BreadcrumbHeader';
-import {actions, defaultBreadcrumb, defaultBreadcrumbLongTitle, defaultTabs} from './ExamplesUtils';
-
-export class BreadcrumbHeaderExample extends React.Component<any, any> {
-    render() {
-        return (
-            <div className="mt2">
-                <div className="form-group">
-                    <label className="form-control-label">Default basic header</label>
-                    <div className="form-control">
-                        <BreadcrumbHeader breadcrumb={defaultBreadcrumb} />
-                    </div>
-                </div>
-
-                <div className="form-group">
-                    <label className="form-control-label">Breadcrumb header with description</label>
-                    <div className="form-control">
-                        <BreadcrumbHeader
-                            breadcrumb={defaultBreadcrumb}
-                            description="Simple description for the title"
-                        />
-                    </div>
-                </div>
-                <div className="form-group">
-                    <label className="form-control-label">Breadcrumb header with long description</label>
-                    <div className="form-control">
-                        <BreadcrumbHeader
-                            breadcrumb={defaultBreadcrumb}
-                            description="Simple description Simple description Simple description Simple description Simple description Simple description Simple description Simple description Simple description Simple description Simple description"
-                        />
-                    </div>
-                </div>
-                <div className="form-group">
-                    <label className="form-control-label">Breadcrumb header without border bottom</label>
-                    <div className="form-control">
-                        <BreadcrumbHeader
-                            breadcrumb={defaultBreadcrumb}
-                            description="Simple description for the title"
-                            hasBorderBottom={false}
-                        />
-                    </div>
-                </div>
-                <div className="form-group">
-                    <label className="form-control-label">Breadcrumb header without padding</label>
-                    <div className="form-control">
-                        <BreadcrumbHeader
-                            breadcrumb={defaultBreadcrumb}
-                            description="Simple description for the title"
-                            hasPadding={false}
-                        />
-                    </div>
-                </div>
-                <div className="form-group">
-                    <label className="form-control-label">Breadcrumb header with actions</label>
-                    <div className="form-control">
-                        <BreadcrumbHeader
-                            breadcrumb={defaultBreadcrumb}
-                            description="Simple description for the title"
-                            actions={actions}
-                        />
-                    </div>
-                </div>
-                <div className="form-group">
-                    <label className="form-control-label">Breadcrumb header with tabs</label>
-                    <div className="form-control">
-                        <BreadcrumbHeader
-                            breadcrumb={defaultBreadcrumbLongTitle}
-                            description="Simple description for the title"
-                            actions={actions}
-                            tabs={defaultTabs}
-                        />
-                    </div>
-                </div>
-                <div className="form-group">
-                    <label className="form-control-label">Breadcrumb with a node as action</label>
-                    <div className="form-control">
-                        <BreadcrumbHeader
-                            breadcrumb={defaultBreadcrumbLongTitle}
-                            description="Simple description for the title"
-                            actions={[
-                                {
-                                    content: (
-                                        <Button>
-                                            <Svg svgName={'add'} svgClass={'icon'} />
-                                        </Button>
-                                    ),
-                                },
-                            ]}
-                        />
-                    </div>
-                </div>
-                <div className="form-group">
-                    <label className="form-control-label">Breadcrumb with a number as action</label>
-                    <div className="form-control">
-                        <BreadcrumbHeader
-                            breadcrumb={defaultBreadcrumbLongTitle}
-                            description="Simple description for the title"
-                            actions={[
-                                {
-                                    content: 1,
-                                },
-                            ]}
-                        />
-                    </div>
-                </div>
-            </div>
-        );
-    }
-}
