+++ packages/react-vapor/src/components/input/InputDescription.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
 
 export const InputDescription: React.FunctionComponent = ({children}) => (
-    <section className="text-dark-medium-grey">{children}</section>
+    <section className="text-dark-medium-grey help-text">{children}</section>
 );
